+++ src/components/RecommendTeamList/index.js
@@ -1,0 +1,5 @@
+function RecommendTeamList() {
+
+}
+
+export default RecommendTeamList;
