+++ src/components/UserBuilding/style.css
@@ -1,12 +1,11 @@
 .userform-box {
   width: 500px;
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  margin-left: -250px;
-  margin-top: -400px;
-  padding: 20px;
+  margin: 0 auto;
+  padding: 50px;
   text-align: center;
+  margin-top: 30px;
+  border-radius: 16px;
+  box-shadow: 0px 0px 30px #eee;
 }
 .userform-title,
 .lang-title,
