--- src/components/UserBuilding2/style.css
@@ -1,22 +1,0 @@
-.userform-box {
-  width: 500px;
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  margin-left: -250px;
-  margin-top: -400px;
-  padding: 20px;
-  text-align: center;
-}
-.userform-title {
-  font-size: 36px;
-  color: #343a40;
-  text-align: center;
-}
-.keyword-box {
-  text-align: center;
-}
-.hide {
-  width: -auto;
-  height: -auto;
-}
