+++ src/components/UserDetail/style.css
@@ -1,0 +1,18 @@
+.userdetail-box{
+  width: 500px;
+  margin: 0 auto;
+  margin-top: 30px;
+  padding: 50px;
+  border-radius: 16px;
+  box-shadow: 0px 0px 30px #eee;
+}
+.userdetail-button{
+  display: flex;
+  justify-content: center;
+}
+.update-button {
+  margin-right: 10px;
+}
+.delete-button {
+  margin-left: 10px;
+}
