+++ src/components/UserList/style.css
@@ -1,0 +1,5 @@
+.userlist-box {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+}
