+++ src/pages/PostPage.js
@@ -1,7 +1,6 @@
 import { Routes, Route } from "react-router-dom";
 import TeamBuilding from "../components/TeamBuilding";
 import UserBuilding from "../components/UserBuilding";
-import UserBuilding2 from "../components/UserBuilding2";
 import EditUser from "../components/EditUser";
 import UserDetail from "../components/UserDetail";
 
@@ -11,7 +10,6 @@ function PostPage() {
       <Routes>
         <Route path="team" element={<TeamBuilding />}></Route>
         <Route path="user" element={<UserBuilding />}></Route>
-        <Route path="usertest" element={<UserBuilding2 />}></Route>
         <Route path="user/:postId" element={<UserDetail />}></Route>
         <Route path="user/update/:postId" element={<EditUser/>}></Route>
       </Routes>
